Renaming, bug fixes, JSDoc.
