+++ packages/workbox-range-requests/README.md
@@ -1,16 +1,16 @@
-# workbox-cache-range-response
+# workbox-range-requests
 
 This library creates a new Response, given a source Response and a Range header value.
 
 ## Installation
 
 ```sh
-npm install --save-dev workbox-cache-range-response
+npm install --save-dev workbox-range-requests
 ```
 
 ## Documentation
 
-Read more at this module's [documentation page](https://workboxjs.org/reference-docs/latest/module-workbox-cache-range-response.html).
+Read more at this module's [documentation page](https://workboxjs.org/reference-docs/latest/module-workbox-range-requests.html).
 
 # What's Workbox?
