+++ packages/workbox-range-requests/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "workbox-cache-range-response",
+  "name": "workbox-range-requests",
   "version": "1.1.0",
   "description": "This library creates a new Response, given a source Response and a Range header value.",
   "keywords": [
@@ -16,13 +16,13 @@
     "build"
   ],
   "scripts": {
-    "prepublish": "gulp build --project workbox-cache-range-response"
+    "prepublish": "gulp build --project workbox-range-requests"
   },
   "author": "Google's Web DevRel Team",
   "license": "Apache-2.0",
   "repository": "googlechrome/workbox",
   "bugs": "https://github.com/googlechrome/workbox/issues",
-  "homepage": "https://github.com/GoogleChrome/workbox/tree/master/packages/workbox-cache-range-response",
-  "main": "build/importScripts/workbox-cache-range-response.prod.v1.1.0.js",
-  "module": "build/modules/workbox-cache-range-response.prod.v1.1.0.mjs"
+  "homepage": "https://github.com/GoogleChrome/workbox/tree/master/packages/workbox-range-requests",
+  "main": "build/importScripts/workbox-range-requests.prod.v1.1.0.js",
+  "module": "build/modules/workbox-range-requests.prod.v1.1.0.mjs"
 }
